% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gordi_plot.R
\name{gordi_plot}
\alias{gordi_plot}
\title{Passes object created by gordi to edits made by ggplot2 functions}
\usage{
gordi_plot(
  pass,
  filename = NULL,
  customize = TRUE,
  width = 8,
  height = 6,
  units = "in",
  dpi = 300,
  device = NULL
)
}
\arguments{
\item{pass}{A list object produced by [gordi_read()]}

\item{filename}{Optional path to save your plot (e.g. "plot.png"), similar to [ggplot2::ggsave()]. If NULL, no file is created.}

\item{customize}{Logical; if TRUE prints a short message about customization.}

\item{width}{Integer; passed to [ggplot2::ggsave()] when `filename` is provided. Used to customize plot size expressed by `units`  argument when saving, as in [ggplot2::ggsave()].}

\item{height}{Integer; passed to [ggplot2::ggsave()] when `filename` is provided. Used to customize plot size expressed by `units` argument when saving, as in [ggplot2::ggsave()].}

\item{units}{String; allowed units: 'in', 'cm', 'mm', 'px', in which the `width` and `height` arguments are expressed. Passed to [ggplot2::ggsave()] when `filename` is provided.}

\item{dpi}{Integer; passed to [ggplot2::ggsave()] when `filename` is provided. Used to customize plot resolution when saving, as in [ggplot2::ggsave()].}

\item{device}{Device to use. Similar to `device` argument in [ggplot2::ggsave()]. If NULL (default), the device is guessed based on the `filename` extension.}
}
\value{
A ggplot object (even when saving to filename).
}
\description{
The function [gordi_plot()] returns the ggplot object created by functions:[gordi_sites()] and/or [gordi_species()] after [gordi_read()].
}
\details{
You can either save the plot via `filename` (wrapper over [ggplot2::ggsave()]), or just get the ggplot object to customize it further with ggplot syntax (e.g. `+ theme(...)`, etc.).
}
\examples{
# get a sites ggplot object and further customize
p <- gordi_read(m, dune.env)|> gordi_sites()|> gordi_plot()
p + theme_minimal()
# customization of sites ggplot object in one line
gordi_read(m, dune.env)|> gordi_sites()|> gordi_plot()+ theme_minimal()
# saving species ggplot object as png file in the directory of the current R-project (still returns the ggplot object)
gordi_read(m, dune.env)|> gordi_species()|> gordi_plot(filename = 'plot.png')

}
\seealso{
[gordi_read()], [gordi_sites()], [gordi_species()], [ggplot2::ggsave()]
}
