% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gordi_read.R
\name{gordi_read}
\alias{gordi_read}
\title{Read ordination output into a standardized structure}
\usage{
gordi_read(
  m,
  env = NULL,
  traits = NULL,
  choices = 1:2,
  scaling = "symm",
  correlation = F,
  hill = F
)
}
\arguments{
\item{m}{An ordination result: object of class `rda`, `cca`, `capscale`,
`decorana`, or `metaMDS`}

\item{env}{A data frame with environmental variables of samples; must have
the same number of samples (rows) as the species table}

\item{traits}{A data frame of species traits; one species can have only one
trait value in its dedicated column}

\item{choices}{Numeric vector of selected ordination axes; defaults to the
first two axes}

\item{scaling}{Scaling of species and site scores; see [vegan::scores.cca()]}

\item{correlation}{Logical; default = FALSE, see [vegan::scores.cca()] or 
[vegan::scores.rda()] for details}

\item{hill}{Logical; default = FALSE, see [vegan::scores.cca()] for details}
}
\value{
A list with elements:
  \describe{
    \item{m}{the original ordination object}
    \item{explained_variation}{eigenvalues divided by total inertia, or NA for DCA/NMDS}
    \item{site_scores}{tibble of site (sample) scores}
    \item{species_scores}{tibble of species scores}
    \item{predictor_scores}{tibble of biplot scores for environmental predictors}
    \item{env}{environmental data frame (if supplied)}
    \item{traits}{trait data frame (if supplied)}
    \item{choices}{selected axes}
    \item{type}{character string with ordination type}
    \item{axis_names}{axis labels}
    \item{species_names}{species names}
    \item{predictor_names}{predictor names, or NA for DCA/NMDS}
  }
}
\description{
`gordi_read` is the first function in the family of `gordi_`functions. 
It takes the output of `rda`, `cca`, `capscale`, `decorana` and `metaMDS` and 
translates in into a format usable by other `gordi_` functions. `gordi_read` 
was not tested for partial ordination (with `Condition()`) yet, and axis labels
are the default ones from vegan.
}
\examples{
library(vegan)
library(tidyverse)
data(dune)
m <- capscale(dune ~ 1)
o <- gordi_read(m)
o
}
\seealso{
[gordi_sites()], [gordi_read()], [gordi_predict()], [gordi_species()]
}
